Fix `recover` output
This commit fixes the output of `rms recover`. Previously the script would not
tell if the file could be recovered or was not recovered because the line

```
mv /tmp/rms/$1 . || echo -e "$CROSS Unable to move file!" && exit
echo -e "$CHECKSIGN Recovered file to $(pwd)"
```

would always exit if the file could be moved but would not run the next line.
This has been a programming mistake and logical issue. It has now been converted
to a if-else block.